Cinematic protester prefab
